(svn r5974) -Codechange: added casts all around the place to make Windows 64bit happy (michi_cc)

+++ bmp.c
@@ -15,7 +15,7 @@ void BmpInitializeBuffer(BmpBuffer *buffer, FILE *file) {
 
 static inline void AdvanceBuffer(BmpBuffer *buffer)
 {
-	buffer->read = fread(buffer->data, 1, BMP_BUFFER_SIZE, buffer->file);
+	buffer->read = (int)fread(buffer->data, 1, BMP_BUFFER_SIZE, buffer->file);
 	buffer->pos  = 0;
 }
